requirements.txt
pytest
PyJWT>=2.8.0
flask-cors>=4.0.0